Backend/test_back_end/requirements.txt
altgraph
amqp
asgiref
attrs

backports.zoneinfo

contourpy
coreapi
coreschema
cryptography

daphne
distlib
Django
django-cors-headers
django-cors-middleware
django-enumfield
django-mssql-backend
django-pyodbc
django-rest-enumfield
django-timezone-field
djangorestframework
future
Jinja2
kiwisolver
kombu
ldap3
MarkupSafe
matplotlib
mssql-django

packaging



platformdirs
prompt-toolkit
psycopg2
psycopg2-binary
pyasn1
pyasn1-modules
pycparser
pyinstaller
pyinstaller-hooks-contrib
PyJWT
PyNaCl
pyodbc
PyYAML
requests

scipy
seaborn
service-identity
simplejson
six
sqlparse
telegram
tenacity
txaio
typing_extensions
tzdata
uritemplate
urllib3
vine
virtualenv
watchdog
wcwidth
zipp
zope.interface